+++ app/src/main/java/com/example/filmfinder/data/repository/DetailsRepository.kt
@@ -4,5 +4,5 @@ import com.example.filmfinder.data.MovieDTO
 import retrofit2.Callback
 
 interface DetailsRepository {
-    fun getMovieFromServer(id:Int,callback: Callback<MovieDTO>)
+    fun getMovieFromServer(id:Long,callback: Callback<MovieDTO>)
 }
